Complete rewrite to make simpler
Instead of using Rc<RefCell<...>> to hold all of the nodes and share
them among many places, like the dictionary, parents, and children,
indices into a vec now represent connections between nodes. This vec is
owned by the tree, and owns all of the nodes. Any time another struct
wants to reference a node, it uses the index into the vec that will
return that node, as a usize. This makes logic and structures easier to
understand, and should speed up the functions dealing with nodes.

Further, both the decode and the encode function have been moved into
impl functions on HuffmanTree. This allows use of every function of the
tree with one use line, and simplifies the calls to these functions.

Tests have been added for more cases.

TODO: return Result<Vec<T>> from the encode and decode functions to
handle errors.